Run at a little higher speed.
It appears that higher speeds require generally smaller PID parameters.
This is reflected in the kinematic motion model

ψ(t+1) = ψ(t) + v/L ∗ δ ∗ dt

where ψ is the yaw angle,
v is the forward velocity,
L is the distance from axel to center of mass,
and δ is the turn angle.

I.e., yaw rate depends on speed.